VOSONDash-package {VOSONDash}	R Documentation

Interface for collection and interactive analysis of social networks

Description

VOSONDash provides functions and an interface in the form of an interactive R Shiny web application for the visualisation and analysis of network data. The app has sections for visualising and manipulating network graphs, performing text analysis, and displaying network metrics. It also has an interface for the collection of social network data using the vosonSML R package.

Author(s)

Bryan Gertzel and Robert Ackland.

[Package VOSONDash version 0.5.7 Index]