applyCategoricalFilters {VOSONDash}	R Documentation

Filter out graph vertices not in selected category

Description

This function removes vertices that are not in the selected categories values list or sub-categories.

Usage

applyCategoricalFilters(
  g,
  selected_cat,
  selected_subcats,
  cat_prefix = "vosonCA_"
)

Arguments

g
igraph graph object.

selected_cat
Character string. Selected vertex category without prefix.

selected_subcats
List. Selected sub-category values to include in graph.

cat_prefix
Character string. Category attribute prefix format to match. Default is "vosonCA_".

Value

An igraph graph object.

Examples

## Not run: 
##D # return a graph containing only vertices that have the vertex category 
##D # attribute "vosonCA_Stance" value "liberal"
##D g <- loadPackageGraph("DividedTheyBlog_40Alist_release.graphml")
##D 
##D g <- applyCategoricalFilters(g, "Stance", c("liberal"))
## End(Not run)

[Package VOSONDash version 0.5.7 Index]