applyComponentFilter {VOSONDash}	R Documentation

Filter out graph vertices not in component size range

Description

This function removes any graph vertices that are in components that fall outside of the specified component size range.

Usage

applyComponentFilter(g, component_type = "strong", component_range)

Arguments

g
igraph graph object.

component_type
Character string. Use strongly or weakly connected components by specifying "strong" or "weak". Ignored for undirected graphs. Default is "strong".

component_range
Numeric vector. Min and max values or size range of component.

Value

An igraph graph object.

[Package VOSONDash version 0.5.7 Index]