applyGraphFilters {VOSONDash}	R Documentation

Filter out graph vertices and edges from graph object that are isolates, multi edge or edge loops

Description

This function removes isolate vertices, multiple edges between vertices and or vertex edge loops from a graph.

Usage

applyGraphFilters(g, isolates = TRUE, multi_edge = TRUE, loops_edge = TRUE)

Arguments

g
igraph graph object.

isolates
Logical. Include isolate vertices in graph. Default is TRUE.

multi_edge
Logical. Include multiple edges between vertices in graph. Default is TRUE.

loops_edge
Logical. Include vertex edge loops in graph. Default is TRUE.

Value

An igraph graph object.

Note

Removing multiple edges or edge loops from a graph will simplify it and remove other edge attributes.

[Package VOSONDash version 0.5.7 Index]