collectTwitterData {VOSONDash}	R Documentation

Collect twitter data

Description

This function is a wrapper for collecting tweets using vosonSML::Collect.

Usage

collectTwitterData(
  cred,
  search_term,
  search_type,
  tweet_count,
  include_retweets,
  retry_on_rate_limit,
  language,
  date_until,
  since_id,
  max_id
)

Arguments

cred
vosonSML twitter credential object.

search_term
Character string. Twitter search term.

search_type
Character string. Search type "mixed", "recent" or "popular".

tweet_count
Numeric. Number of tweets to collect.

include_retweets
Logical. Include retweets in the results.

retry_on_rate_limit
Logical. Wait and retry when the twitter api rate limit is reached.

language
Character string. Language code of tweets to collect as two character ISO 639-1 code.

date_until
Character string. Date to collect tweets to in format "YYYY-MM-DD".

since_id
Numeric. Collect tweets with a tweet id more recent than since_id.

max_id
Numeric. Collect tweets with a tweet id older than max_id.

Value

A vosonSML twitter dataframe.

[Package VOSONDash version 0.5.7 Index]