createTwitterDevToken {VOSONDash}	R Documentation

Create an auth token with twitter app dev keys

Description

This function is a wrapper for vosonSML::Authenticate with twitter app developer keys. The properties type and created are added to the credential object to assist with VOSONDash token management.

Usage

createTwitterDevToken(app_name, keys)

Arguments

app_name
Character string. Twitter app name.

keys
List. Named list of twitter app API keys.

Value

A vosonSML twitter credential object.

[Package VOSONDash version 0.5.7 Index]