emptyPlotMessage {VOSONDash}	R Documentation

Create an empty plot with text

Description

This function creates an empty plot that can be used to display a text message. Intended to be used in a series of plots to indicate that an individual plot cannot be created for some reason and still maintain a plot aesthetic.

Usage

emptyPlotMessage(message = "No plot available.")

Arguments

message
Character string. Text message to centre on empty plot. Default text is "No plot available.".

Value

An empty plot with text message.

[Package VOSONDash version 0.5.7 Index]