getNetworkMetrics {VOSONDash}	R Documentation

Get graph network metrics

Description

Function creates a vector of calculated network metrics for a graph.

Usage

getNetworkMetrics(g, component_type = "strong")

Arguments

g
igraph graph object.

component_type
Character string. Use strongly or weakly connected components by specifying "strong" or "weak". Ignored for undirected graphs. Default is "strong".

Value

Network metrics as named vector.

[Package VOSONDash version 0.5.7 Index]