getVertexCategories {VOSONDash}	R Documentation

Get a list of vertex category attribute names and values

Description

This function returns a list of graph vertex attribute names that match a category attribute prefix format and their unique values.

Usage

getVertexCategories(g, cat_prefix = "vosonCA_")

Arguments

g
igraph graph object.

cat_prefix
Character string. Category attribute prefix format to match. Default is "vosonCA_".

Value

A named list of vertex category attributes and values.

Examples

## Not run: 
##D # get a list of voson vertex categories and values
##D g <- loadPackageGraph("DividedTheyBlog_40Alist_release.graphml")
##D 
##D vcats <- getVertexCategories(g)
##D 
##D # vcats
##D # $Stance
##D # [1] "conservative" "liberal"  
## End(Not run)

[Package VOSONDash version 0.5.7 Index]