Get reddit thread listings for one or more subreddits

This function is a wrapper for collecting reddit subreddit listings using vosonSML::Collect.

get_rddt_lst_data(
  subreddits = NULL,
  sort = "new",
  period = "all",
  max = 25,
  ...
)

Arguments

subreddits

Character vector. Subreddit urls to collect threads from.

...

Arguments passed on to vosonSML::collect_reddit_listings

sort

Character vector. Listing thread sort order. Options are "hot", "top", "new", and "rising". Default is "hot".

period

Character vector. Listing top threads by time period. Only applicable to sort order by "top". Options are "hour", "day", "week", "month", "year" and "all". Default is "all".

max

Numeric vector. Maximum number of threads in listing to return. Default is 25.

waitTime

Numeric vector. Time range in seconds to select random wait from in-between url collection requests. Minimum is 3 seconds. Default is c(6, 8) for a wait time chosen from between 6 and 8 seconds.

ua

Character string. Override User-Agent string to use in Reddit thread requests. Default is option("HTTPUserAgent") value as set by vosonSML.

Value

A vosonSML subreddit listing dataframe.