logMessage {VOSONDash}	R Documentation

Add message to log queue

Description

This function adds a text message to a queue or list with a count limiting how many messages are stored. The queue stores count messages based on first in first out.

Usage

logMessage(messages, add_message, txt = FALSE, count = 20)

Arguments

messages
Character vector. Lines or text log messages.

add_message
Character string. Text log message to add to messages

txt
Logical. Return messages as single character string delimited by newline characters.

count
Numeric. Return queue of count messages and discard the rest.

Value

Messages as vector or character string.

[Package VOSONDash version 0.5.7 Index]