systemTimeFilename {VOSONDash}	R Documentation

Create a file name with system date time prefix

Description

This function uses the system date and time to create a unique file name.

Usage

systemTimeFilename(name_suffix, name_ext, clean = FALSE)

Arguments

name_suffix
Character string. Name part of file name to append to date time part.

name_ext
Character string. File extension without the period. For example, "txt".

clean
Logical. Remove problematic file system characters from file name part. Default is FALSE.

Value

A unique date time file name as character string.

[Package VOSONDash version 0.5.7 Index]