wordCloudPlot {VOSONDash}	R Documentation

Create a wordcloud plot

Description

This function creates a wordcloud plot from word frequencies.

Usage

wordCloudPlot(
  word_freqs,
  seed = NULL,
  min_freq = 1,
  max_words = 50,
  pcolors = NULL,
  family = NULL,
  ...
)

Arguments

word_freqs
Table. Table of word frequencies.

seed
Numeric. Seed value can be supplied to reproduce a word cloud layout.

min_freq
Numeric. Minimum word frequency to include a word in the word cloud. Default is 1.

max_words
Numeric. Maximum number of words to render in the word cloud. Default is 50.

pcolors
List. Colors to assign categorical variable in the plot or palette to use if random.color. Default is NULL.

family
Character. Set a font family for plot labels. Default is NULL.

...
Arguments passed on to wordcloud::wordcloud

random.order

plot words in random order. If false, they will be plotted in decreasing frequency

random.color

choose colors randomly from the colors. If false, the color is chosen based on the frequency

rot.per

proportion words with 90 degree rotation

Value

A wordcloud plot.

[Package VOSONDash version 0.5.7 Index]