wordFreqChart {VOSONDash}	R Documentation

Create a word frequency chart

Description

This function creates a horizontal barchart of word frequencies.

Usage

wordFreqChart(
  word_freqs,
  min_freq = 1,
  top_count = 20,
  pcolors = NULL,
  family = NULL
)

Arguments

word_freqs
Dataframe. Word frequencies.

min_freq
Numeric. Minimum frequency for a word to be included in the chart. Default is 1.

top_count
Numeric. Top count of words to render in word frequency chart. Default is 20.

pcolors
List. Colors to assign categorical variable in the plot. Default is NULL.

family
Character string. Set a font family for plot labels. Default is NULL.

Value

A barchart plot.

[Package VOSONDash version 0.5.7 Index]