wordSentChart {VOSONDash}	R Documentation

Create an NRC emotion chart

Description

This function creates a horizontal barchart measuring and sorting the eight NRC lexicon emotions. Emotions are measured as the proportion of the total value of the eight emotions in the text as a percentage.

Usage

wordSentChart(data, pcolors = NULL)

Arguments

data
Dataframe. NRC emotions table.

pcolors
List. Colors to assign categorical variable in the plot. Default is NULL.

Value

A barchart plot.

Note

Uses the syuzhet package implementation of Saif Mohammad’s NRC Emotion lexicon.

[Package VOSONDash version 0.5.7 Index]