wordSentValenceChart {VOSONDash}	R Documentation

Create an NRC sentiment valence chart

Description

This function creates a vertical barchart of the sum of negative and positive sentiments, and the valence or net sentiment in a text corpus.

Usage

wordSentValenceChart(data)

Arguments

data
Dataframe. NRC emotions table.

Value

A barchart plot.

[Package VOSONDash version 0.5.7 Index]